Feat: Add CellView UI Update
